+++ src/parser/block/paragraph/reference.test.ts
@@ -16,18 +16,18 @@ describe('Unit: parser/block/paragraph/reference', () => {
     });
 
     it('valid', () => {
-      assert.deepStrictEqual(inspect(parser('>0')), [['<a class="reference" rel="noopener">&gt;0</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>a')), [['<a class="reference" rel="noopener">&gt;a</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>A')), [['<a class="reference" rel="noopener">&gt;A</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>a ')), [['<a class="reference" rel="noopener">&gt;a</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>a\n')), [['<a class="reference" rel="noopener">&gt;a</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>\\')), [['<a class="reference" rel="noopener">&gt;\\</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>\\ ')), [['<a class="reference" rel="noopener">&gt;\\</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>\\\n')), [['<a class="reference" rel="noopener">&gt;\\</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>>0')), [['<a class="reference" rel="noopener">&gt;&gt;0</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>0\n>1')), [['<a class="reference" rel="noopener">&gt;0</a>', '<br>', '<a class="reference" rel="noopener">&gt;1</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>0\n>>1')), [['<a class="reference" rel="noopener">&gt;0</a>', '<br>', '<a class="reference" rel="noopener">&gt;&gt;1</a>', '<br>'], '']);
-      assert.deepStrictEqual(inspect(parser('>>0\n>1')), [['<a class="reference" rel="noopener">&gt;&gt;0</a>', '<br>', '<a class="reference" rel="noopener">&gt;1</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>0')), [['<a class="reference" rel="noopener" data-level="1">&gt;0</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>a')), [['<a class="reference" rel="noopener" data-level="1">&gt;a</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>A')), [['<a class="reference" rel="noopener" data-level="1">&gt;A</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>a ')), [['<a class="reference" rel="noopener" data-level="1">&gt;a</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>a\n')), [['<a class="reference" rel="noopener" data-level="1">&gt;a</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>\\')), [['<a class="reference" rel="noopener" data-level="1">&gt;\\</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>\\ ')), [['<a class="reference" rel="noopener" data-level="1">&gt;\\</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>\\\n')), [['<a class="reference" rel="noopener" data-level="1">&gt;\\</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>0')), [['<a class="reference" rel="noopener" data-level="2">&gt;&gt;0</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>0\n>1')), [['<a class="reference" rel="noopener" data-level="1">&gt;0</a>', '<br>', '<a class="reference" rel="noopener" data-level="1">&gt;1</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>0\n>>1')), [['<a class="reference" rel="noopener" data-level="1">&gt;0</a>', '<br>', '<a class="reference" rel="noopener" data-level="2">&gt;&gt;1</a>', '<br>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>0\n>1')), [['<a class="reference" rel="noopener" data-level="2">&gt;&gt;0</a>', '<br>', '<a class="reference" rel="noopener" data-level="1">&gt;1</a>', '<br>'], '']);
     });
 
   });